# examples/output-sqlite/openapi.yaml
...
openapi: 3.0.3
info:
  title: InventoryPlatform
  version: 1.0.0
  description: Generated by MarrowScript compiler
servers:
  - 
    url: "http://localhost:3000"
paths:
...
  /items:
    get:
      summary: List Item
      operationId: listItem
      tags:
        - Item
      parameters:
        - 
          name: page
          in: query
          schema:
            type: integer
            default: 1
        - 
          name: page_size
          in: query
          schema:
            type: integer
            default: 50
      responses:
        200:
          description: List of Item
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Item"
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer
        401:
          description: Unauthorized
    post:
      summary: Create Item
      operationId: createItem
      tags:
        - Item
      security:
        - 
          BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Item"
      responses:
        200:
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        400:
          description: Bad request
        401:
          description: Unauthorized
        422:
          description: Precondition failed
  /items/{id}:
    get:
      summary: Get Item
      operationId: getItem
      tags:
        - Item
      parameters:
        - 
          name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      security:
        - 
          BearerAuth: []
      responses:
        200:
          description: Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        400:
          description: Not found
        401:
          description: Unauthorized
    put:
      summary: Update Item
      operationId: updateItem
      tags:
        - Item
      parameters:
        - 
          name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      security:
        - 
          BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Item"
      responses:
        200:
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        400:
          description: Bad request
        401:
          description: Unauthorized
        422:
          description: Precondition failed
    delete:
      summary: Delete Item
      operationId: deleteItem
      tags:
        - Item
      parameters:
        - 
          name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      security:
        - 
          BearerAuth: []
      responses:
        200:
          description: Deleted
        400:
          description: Not found
        401:
          description: Unauthorized
  /items/consume-item:
    post:
      summary: consume_item on Item
      operationId: consume_itemItem
      tags:
        - Item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                player_id:
                  type: string
                  format: uuid
                  description: UUID of an existing Player row (alternative to inline body)
                item_id:
                  type: string
                  format: uuid
                  description: UUID of an existing Item row (alternative to inline body)
                id:
                  type: string
                  format: uuid
                created_at:
                  type: string
                  format: date-time
                updated_at:
                  type: string
                  format: date-time
                name:
                  type: string
                quantity:
                  type: integer
                rarity:
                  type: string
                tradeable:
                  type: boolean
                owner_id:
                  type: string
                  format: uuid
                metadata:
                  type: object
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  action:
                    type: string
        400:
          description: Bad request
        401:
          description: Unauthorized
        404:
          description: Not found
        422:
          description: Precondition failed
      security:
        - 
          BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Item:
      type: object
      properties:
        id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        name:
          type: string
        quantity:
          type: integer
        rarity:
          type: string
        tradeable:
          type: boolean
        owner_id:
          type: string
          format: uuid
        metadata:
          type: object
...