# examples/output-sqlite/src/routes/item.ts
// Generated by MarrowScript compiler. DO NOT EDIT.
// Service: ItemService
import { Router, Request, Response } from "express";
import { v4 as uuid } from "uuid";
import { query, queryOne, execute, pool } from "../db";
import { eventBus } from "../events";
import { requireAuth, AuthContext } from "../auth";
import rateLimit from "express-rate-limit";
import { auditLog } from "../audit";
import { logger } from "../logger";
import { counter } from "../metrics";
import { ItemCreateSchema, ItemUpdateSchema } from "../schemas";
import * as __algorithms from "../algorithms";
const { shortestPath, topologicalSort, binarySearch, bipartiteMatching, roundRobin, weightedAverage, percentile, rankBy, consistentHash } = __algorithms as any;

import { transitionItem, ITEM_INITIAL } from "../state_machines/item";

export const itemsRouter = Router();

// Rate limiter from policy declaration
const __routeRateLimit = rateLimit({ windowMs: 60000, max: 100, standardHeaders: true, legacyHeaders: false });

// CREATE
itemsRouter.post("/", __routeRateLimit, requireAuth, async (req: Request, res: Response) => {
  // Validate the request body against the generated Zod schema.
  const __parsed = ItemCreateSchema.safeParse(req.body);
  if (!__parsed.success) {
    return res.status(400).json({ error: { code: "VALIDATION_FAILED", message: "Invalid request body", issues: __parsed.error.flatten() } });
  }
  const __body = __parsed.data;
  try {
    const id = uuid();
    const { name, quantity, rarity, tradeable, owner_id, metadata } = __body as any;
    const state = ITEM_INITIAL;
    const sql = `INSERT INTO items (id, name, quantity, rarity, tradeable, owner_id, metadata, state) VALUES ($1, $2, $3, $4, $5, $6, $7, $8) RETURNING *`;
    const rows = await query(sql, [id, name, quantity, rarity, tradeable, owner_id, metadata, state]);
    counter("entity.created", { entity: "Item" });
    res.status(201).json(rows[0]);
  } catch (e: any) {
    res.status(400).json({ error: { code: "CREATE_FAILED", message: e.message } });
  }
});

// READ
itemsRouter.get("/:id", __routeRateLimit, requireAuth, async (req: Request, res: Response) => {
  try {
    const row = await queryOne(`SELECT * FROM items WHERE id = $1`, [req.params.id]);
    if (!row) return res.status(404).json({ error: { code: "NOT_FOUND", message: "Not found" } });
    res.json(row);
  } catch (e: any) {
    res.status(500).json({ error: { code: "DB_ERROR", message: e.message } });
  }
});

// LIST
itemsRouter.get("/", __routeRateLimit, requireAuth, async (req: Request, res: Response) => {
  try {
    const page = parseInt(req.query.page as string) || 1;
    const pageSize = Math.min(parseInt(req.query.page_size as string) || 50, 100);
    const offset = (page - 1) * pageSize;
    const rows = await query(`SELECT * FROM items ORDER BY created_at DESC LIMIT $1 OFFSET $2`, [pageSize, offset]);
    const [{ count }] = await query(`SELECT COUNT(*) as count FROM items`);
    res.json({ items: rows, total: parseInt(count), page, page_size: pageSize });
  } catch (e: any) {
    res.status(500).json({ error: { code: "DB_ERROR", message: e.message } });
  }
});

// UPDATE
const __itemsUpdatable = new Set<string>(["name","quantity","rarity","tradeable","owner_id","metadata","state"]);
itemsRouter.put("/:id", __routeRateLimit, requireAuth, async (req: Request, res: Response) => {
  // 1. Validate body shape and types via Zod (UpdateSchema is partial).
  const __parsed = ItemUpdateSchema.safeParse(req.body);
  if (!__parsed.success) {
    return res.status(400).json({ error: { code: "VALIDATION_FAILED", message: "Invalid request body", issues: __parsed.error.flatten() } });
  }
  // 2. Reject unknown columns (defense in depth — Zod would already strip them).
  const __unknown = Object.keys(req.body || {}).filter(k => !__itemsUpdatable.has(k));
  if (__unknown.length > 0) {
    return res.status(400).json({ error: { code: "UNKNOWN_FIELDS", message: `Unknown fields: ${__unknown.join(", ")}`, fields: __unknown } });
  }
  // 3. Use the Zod-parsed object as the update set.
  const fields: Record<string, unknown> = __parsed.data as Record<string, unknown>;
  if (Object.keys(fields).length === 0) {
    return res.status(400).json({ error: { code: "NO_FIELDS", message: "No updatable fields supplied" } });
  }
  // State machine enforcement
  if (fields.state !== undefined) {
    const current = await queryOne<{ state: string }>(`SELECT state FROM items WHERE id = $1`, [req.params.id]);
    if (!current) return res.status(404).json({ error: { code: "NOT_FOUND", message: "Not found" } });
    // Find the trigger for this state transition
    const trigger = `${current.state}_to_${fields.state}`;
    const tr = transitionItem(current.state as any, trigger);
    if (!tr.ok) return res.status(422).json({ error: { code: "INVALID_TRANSITION", message: `Cannot transition from ${current.state} to ${fields.state}` } });
  }
  const keys = Object.keys(fields);
  const sets = keys.map((k, i) => `${k} = $${i + 2}`).join(", ");
  const values = keys.map(k => fields[k]);
  const sql = `UPDATE items SET ${sets}, updated_at = NOW() WHERE id = $1 RETURNING *`;
  const rows = await query(sql, [req.params.id, ...values]);
  if (rows.length === 0) return res.status(404).json({ error: { code: "NOT_FOUND", message: "Not found" } });
  res.json(rows[0]);
});

// DELETE
itemsRouter.delete("/:id", __routeRateLimit, requireAuth, async (req: Request, res: Response) => {
  try {
    const count = await execute(`DELETE FROM items WHERE id = $1`, [req.params.id]);
    if (count === 0) return res.status(404).json({ error: { code: "NOT_FOUND", message: "Not found" } });
    res.status(204).send();
  } catch (e: any) {
    res.status(500).json({ error: { code: "DB_ERROR", message: e.message } });
  }
});

// CAPABILITY: consume_item [transactional]
itemsRouter.post("/consume-item", __routeRateLimit, requireAuth, auditLog("consume_item", "Item"), async (req: Request, res: Response) => {
  const auth: AuthContext = (req as any).auth;
  const __client = await pool.connect();
  try {
    await __client.query("BEGIN");
    // Fetch entities
    const player = await queryOne(`SELECT * FROM players WHERE id = $1`, [req.body.player_id || req.params.id]);
    if (!player) {
      return res.status(404).json({ error: { code: "NOT_FOUND", message: "player not found" } });
    }
    const item = await queryOne(`SELECT * FROM items WHERE id = $1`, [req.body.item_id || req.params.id]);
    if (!item) {
      return res.status(404).json({ error: { code: "NOT_FOUND", message: "item not found" } });
    }

    // Preconditions
    if (!(player?.state === "active")) {
      return res.status(422).json({ error: { code: "PRECONDITION_FAILED", message: "player.state == \\\"active\\\"" } });
    }
    if (!(item?.owner_id === player?.id)) {
      return res.status(422).json({ error: { code: "PRECONDITION_FAILED", message: "item.owner_id == player.id" } });
    }
    if (!(item?.state === "available")) {
      return res.status(422).json({ error: { code: "PRECONDITION_FAILED", message: "item.state == \\\"available\\\"" } });
    }
    if (!(item?.quantity > 0)) {
      return res.status(422).json({ error: { code: "PRECONDITION_FAILED", message: "item.quantity > 0" } });
    }

    // Effects (applied in declaration order)
    const __effect_0 = await query(`UPDATE items SET quantity = quantity - $1, updated_at = NOW() WHERE id = $2 RETURNING *`, [1, req.body.item_id || req.params.id]);
    if (!__effect_0 || __effect_0.length === 0) {
      throw new Error("Effect failed: item.quantity -= 1");
    }

    res.json({ ok: true, action: "consume_item", entity: player });
    await __client.query("COMMIT");
  } catch (e: any) {
    await __client.query("ROLLBACK");
    if (!res.headersSent) {
      res.status(400).json({ error: { code: "CAPABILITY_FAILED", message: e.message } });
    } else {
      logger.error("late_error_after_response", { event: "consume_item", status: "rejected", metadata: { error: String(e?.message || e).slice(0, 256) } });
    }
  } finally {
    __client.release();
  }
});