# examples/output-sqlite/src/schemas.ts
// Generated by MarrowScript compiler.
import { z } from "zod";

export const PlayerSchema = z.object({
  id: z.string().uuid(),
  username: z.string(),
  email: z.string(),
  password_hash: z.string(),
  score: z.number().int().nonnegative(),
  level: z.number().int().nonnegative(),
  state: z.string(),
  created_at: z.string().datetime(),
  updated_at: z.string().datetime(),
});
export type Player = z.infer<typeof PlayerSchema>;
export const PlayerCreateSchema = PlayerSchema.omit({ id: true, created_at: true, updated_at: true }).partial({ state: true });
export const PlayerUpdateSchema = PlayerSchema.omit({ id: true, created_at: true, updated_at: true }).partial();

export const ItemSchema = z.object({
  id: z.string().uuid(),
  name: z.string(),
  quantity: z.number().int().nonnegative(),
  rarity: z.string(),
  tradeable: z.boolean(),
  owner_id: z.string().uuid(),
  state: z.string(),
  metadata: z.unknown(),
  created_at: z.string().datetime(),
  updated_at: z.string().datetime(),
});
export type Item = z.infer<typeof ItemSchema>;
export const ItemCreateSchema = ItemSchema.omit({ id: true, created_at: true, updated_at: true }).partial({ state: true });
export const ItemUpdateSchema = ItemSchema.omit({ id: true, created_at: true, updated_at: true }).partial();

export const TradeOfferSchema = z.object({
  id: z.string().uuid(),
  created_at: z.string().datetime(),
  updated_at: z.string().datetime(),
  from_player: z.string().uuid(),
  to_player: z.string().uuid(),
  offered_items: z.array(z.string().uuid()),
  requested_items: z.array(z.string().uuid()),
  status: z.string(),
  expires_at: z.string().datetime(),
});
export type TradeOffer = z.infer<typeof TradeOfferSchema>;
export const TradeOfferCreateSchema = TradeOfferSchema.omit({ id: true, created_at: true, updated_at: true });
export const TradeOfferUpdateSchema = TradeOfferSchema.omit({ id: true, created_at: true, updated_at: true }).partial();

export const PlayerServiceCreateInputSchema = z.object({
  username: z.string(),
  email: z.string(),
  score: z.number().int().nonnegative(),
  level: z.number().int().nonnegative(),
  inventory: z.array(z.unknown()),
});

export const PlayerServiceReadInputSchema = z.object({
  id: z.string().uuid(),
});

export const PlayerServiceUpdateInputSchema = z.object({
  username: z.string(),
  email: z.string(),
  score: z.number().int().nonnegative(),
  level: z.number().int().nonnegative(),
  inventory: z.array(z.unknown()),
});

export const PlayerServiceDeleteInputSchema = z.object({
  id: z.string().uuid(),
});

export const PlayerServiceListInputSchema = z.object({
  page: z.number().int().nonnegative(),
  page_size: z.number().int().nonnegative(),
});

export const PlayerServiceCreateTradeInputSchema = z.object({
  from: z.unknown(),
  to: z.unknown(),
  items: z.array(z.unknown()),
});

export const PlayerServiceAcceptTradeInputSchema = z.object({
  trade: z.unknown(),
  acceptor: z.unknown(),
});

export const PlayerServiceAwardXpInputSchema = z.object({
  player: z.unknown(),
  amount: z.number().int().nonnegative(),
});

export const PlayerServiceConsumeItemInputSchema = z.object({
  player: z.unknown(),
  item: z.unknown(),
});

export const ItemServiceCreateInputSchema = z.object({
  name: z.string(),
  quantity: z.number().int().nonnegative(),
  rarity: z.string(),
  tradeable: z.boolean(),
  owner_id: z.string().uuid(),
  metadata: z.unknown(),
});

export const ItemServiceReadInputSchema = z.object({
  id: z.string().uuid(),
});

export const ItemServiceUpdateInputSchema = z.object({
  name: z.string(),
  quantity: z.number().int().nonnegative(),
  rarity: z.string(),
  tradeable: z.boolean(),
  owner_id: z.string().uuid(),
  metadata: z.unknown(),
});

export const ItemServiceDeleteInputSchema = z.object({
  id: z.string().uuid(),
});

export const ItemServiceListInputSchema = z.object({
  page: z.number().int().nonnegative(),
  page_size: z.number().int().nonnegative(),
});

export const ItemServiceConsumeItemInputSchema = z.object({
  player: z.unknown(),
  item: z.unknown(),
});

export const TradeOfferServiceCreateInputSchema = z.object({
  from_player: z.string().uuid(),
  to_player: z.string().uuid(),
  offered_items: z.array(z.string().uuid()),
  requested_items: z.array(z.string().uuid()),
  status: z.string(),
  expires_at: z.string().datetime(),
});

export const TradeOfferServiceReadInputSchema = z.object({
  id: z.string().uuid(),
});

export const TradeOfferServiceUpdateInputSchema = z.object({
  from_player: z.string().uuid(),
  to_player: z.string().uuid(),
  offered_items: z.array(z.string().uuid()),
  requested_items: z.array(z.string().uuid()),
  status: z.string(),
  expires_at: z.string().datetime(),
});

export const TradeOfferServiceDeleteInputSchema = z.object({
  id: z.string().uuid(),
});

export const TradeOfferServiceListInputSchema = z.object({
  page: z.number().int().nonnegative(),
  page_size: z.number().int().nonnegative(),
});

export const TradeOfferServiceAcceptTradeInputSchema = z.object({
  trade: z.unknown(),
  acceptor: z.unknown(),
});

export const GameLobbySubscribeInputSchema = z.object({
  topic: z.string(),
});

export const GameLobbyPublishInputSchema = z.object({
  message: z.unknown(),
});

export const TradeNotificationsSubscribeInputSchema = z.object({
  topic: z.string(),
});

export const TradeNotificationsPublishInputSchema = z.object({
  message: z.unknown(),
});